Feat: user_detail 함수 수정 + GET,PUT,DELETE 한 번에 처리

+++ django/week04_hw/accounts/views.py
@@ -17,39 +17,28 @@ def signup(request):
         users=User.objects.all()
         serializer=UserSerializer(users,many=True)
         return Response(serializer.data,status=status.HTTP_200_OK)
-    # 처음에 POST 만 추가 -> 회원가입을 한 전체 아이디가 보이지 않고, 가장 최신 아이디만 보임 -> GET 추가 
-#회원정보 조회(Read) ->/users/<id>
-@api_view(['GET'])
-def get_user(request, user_id):
+    # 처음에 POST 만 추가 -> 회원가입을 한 전체 아이디가 보이지 않고, 가장 최신 아이디만 보임 -> GET 추가
+     
+#회원 상세보기, 수정, 삭제 -> /users/<int:user_id>/ -> user_detail 함수로 GET, PUT, DELETE 합치기
+
+@api_view(['GET', 'PUT', 'DELETE'])
+def user_detail(request, user_id):
     try:
         user = User.objects.get(pk=user_id)
     except User.DoesNotExist:
         return Response({'error': 'User not found'}, status=status.HTTP_404_NOT_FOUND)
 
-    serializer = UserSerializer(user)
-    return Response(serializer.data,status=status.HTTP_200_OK)
-
-#회원정보 수정(Update) -> /users/<id>
-@api_view(['PUT'])
-def update_user(request, user_id):
-    try:
-        user = User.objects.get(pk=user_id)
-    except User.DoesNotExist: #사용자가 존재하지 않을때
-        return Response({'error': 'User not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    serializer = UserSerializer(user, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
+    if request.method == 'GET':
+        serializer = UserSerializer(user)
         return Response(serializer.data, status=status.HTTP_200_OK)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-#회원정보 삭제(Delete) -> /users/<id>
-@api_view(['DELETE'])
-def delete_user(request, user_id):
-    try:
-        user = User.objects.get(pk=user_id)
-    except User.DoesNotExist: #사용자가 존재하지 않을때
-        return Response({'error': 'User not found'}, status=status.HTTP_404_NOT_FOUND)
+    elif request.method == 'PUT':
+        serializer = UserSerializer(user, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    user.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    elif request.method == 'DELETE':
+        user.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
